Convert prime spec to playground
