+++ SwiftCheck.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,9 @@
 <Workspace
    version = "1.0">
    <FileRef
+      location = "group:SwiftCheck/SwiftCheck.playground">
+   </FileRef>
+   <FileRef
       location = "group:SwiftCheck-iOS.xcodeproj">
    </FileRef>
    <FileRef
