+++ SwiftCheck/SwiftCheck.playground/Sources/SupportCode.swift
@@ -1,0 +1,3 @@
+//
+// This file (and all other Swift source files in the Sources directory of this playground) will be precompiled into a framework which is automatically made available to SwiftCheck.playground.
+//
