+++ SwiftCheck/SwiftCheck.playground/contents.xcplayground
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<playground version='5.0' target-platform='osx' display-mode='rendered'>
+    <timeline fileName='timeline.xctimeline'/>
+</playground>
