+++ SwiftCheck/SwiftCheck.playground/playground.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:../../Desktop/SwiftCheck/SwiftCheck/SwiftCheck.playground">
+   </FileRef>
+</Workspace>
